Pet_Adoption_App/backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pymongo
python-dotenv
python-multipart